refactor: post 명명을 posting 으로 수정
