+++ src/main/java/com/next/genshinflow/application/posting/controller/publicapi/PostingController.java
@@ -1,6 +1,6 @@
-package com.next.genshinflow.application.post.controller.publicapi;
+package com.next.genshinflow.application.posting.controller.publicapi;
 
-import com.next.genshinflow.application.post.response.PostingResponse;
+import com.next.genshinflow.application.posting.response.PostingResponse;
 import com.next.genshinflow.auth.UserAuth;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
