+++ src/main/java/com/next/genshinflow/domain/posting/repository/PostingRepository.java
@@ -1,0 +1,12 @@
+package com.next.genshinflow.domain.posting.repository;
+
+import com.next.genshinflow.domain.posting.entity.PostingEntity;
+import io.swagger.v3.oas.annotations.Hidden;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Hidden
+@Repository
+public interface PostingRepository extends JpaRepository<PostingEntity, Long> {
+
+}
